update to chef 12.5 custom resources
