+++ .rubocop.yml
@@ -1,2 +1,2 @@
 AllCops:
-  TargetChefVersion: 11
+  TargetChefVersion: 12.latest
