+++ libraries/matchers.rb
@@ -1,4 +1,4 @@
-# cookstyle: disable ChefModernize/DefinesChefSpecMatchers
+# cookstyle: disable Chef/Modernize/DefinesChefSpecMatchers
 if defined?(ChefSpec)
   ChefSpec.define_matcher(:chef_client_updater)
